pyproject.toml
[project]
name = "core-api-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.119.1",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
